requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
cssselect2==0.7.0
defusedxml==0.7.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
openai==1.52.2
pillow==11.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
tinycss2==1.4.0
tqdm==4.66.5
typing_extensions==4.12.2
webencodings==0.5.1
